fix(ui): move landing components to src/, restore card links, use centralized text
- Move entidades-carousel and feature-section from root components/ui/ to src/components/pages/landing/
- Wrap FeatureSection card in Link to restore full-card clickability
- Use PAGE_HEADER_TEXT constants in vagas page instead of hardcoded strings
- Use cn() for PageHeader className merging
